1、split models
